requirements_docker.txt
natsort==8.2.0
tqdm==4.64.1
numpy==1.20.0
yt-dlp==2022.9.1
pandas
pytorchvideo==0.1.5
pytorch-lightning==1.7.6
opencv-python==4.6.0.66
Pillow==9.2.0
matplotlib==3.1.2
scipy==1.4.1
scikit-image
albumentations
sk-video==1.1.10
scikit-learn
boto3
seaborn
setuptools==60.10.0